refactor: extracted modules tests from desmosclient.integration.spec.ts
