+++ packages/core/src/desmosclient.integration.spec.ts
@@ -1,65 +1,15 @@
 /* eslint-disable */
-import { fromBase64, fromUtf8, toHex } from "@cosmjs/encoding";
+import { fromBase64, fromUtf8 } from "@cosmjs/encoding";
 import { Profile } from "@desmoslabs/desmjs-types/desmos/profiles/v3/models_profile";
 import { MsgSendEncodeObject, SignerData, StdFee } from "@cosmjs/stargate";
 import { AuthInfo, SignDoc } from "cosmjs-types/cosmos/tx/v1beta1/tx";
-import {
-  Bech32Address,
-  ChainConfig,
-  Proof,
-  SignatureValueType,
-  SingleSignature
-} from "@desmoslabs/desmjs-types/desmos/profiles/v3/models_chain_links";
-import { Any } from "@desmoslabs/desmjs-types/google/protobuf/any";
-import { MsgLinkChainAccount } from "@desmoslabs/desmjs-types/desmos/profiles/v3/msgs_chain_links";
-import { MsgSaveProfile } from "@desmoslabs/desmjs-types/desmos/profiles/v3/msgs_profile";
 import { EncodeObject } from "@cosmjs/proto-signing";
-import Long from "long";
-import { sleep } from "@cosmjs/utils";
 import { DesmosClient } from "./desmosclient";
 import { OfflineSignerAdapter, Signer, SigningMode } from "./signers";
-import {
-  assertTxSuccess,
-  defaultGasPrice,
-  getAminoSignerAndClient,
-  getDirectSignerAndClient,
-  pollTx,
-  TEST_CHAIN_URL,
-  testUser1,
-  testUser2
-} from "./testutils";
-import {
-  getPubKeyBytes,
-  getPubKeyRawBytes,
-  getSignatureBytes,
-  getSignedBytes,
-  SignatureResult
-} from "./signatureresult";
+import { assertTxSuccess, defaultGasPrice, getDirectSignerAndClient, pollTx, testUser1, testUser2 } from "./testutils";
+import { getPubKeyBytes, getPubKeyRawBytes, SignatureResult } from "./signatureresult";
 import { MsgAuthenticateEncodeObject, MsgAuthenticateTypeUrl } from "./modules/desmjs/v1";
-import {
-  bech32AddressToAny,
-  MsgLinkChainAccountEncodeObject,
-  MsgSaveProfileEncodeObject,
-  MsgSaveProfileTypeUrl,
-  singleSignatureToAny
-} from "./modules/profiles/v3";
-import { MsgCreateSubspaceEncodeObject, MsgCreateSubspaceTypeUrl } from "./modules/subspaces/v3";
-import { MsgCreatePostEncodeObject, MsgCreatePostTypeUrl } from "./modules/posts/v3";
-import { DoNotModify } from "./modules/consts";
-import {
-  MsgAddReasonEncodeObject,
-  MsgAddReasonTypeUrl,
-  MsgCreateReportEncodeObject,
-  MsgCreateReportTypeUrl,
-  postTargetToAny
-} from "./modules/reports/v1";
-import {
-  MsgAddReactionEncodeObject,
-  MsgAddReactionTypeUrl,
-  MsgAddRegisteredReactionEncodeObject,
-  MsgAddRegisteredReactionTypeUrl,
-  registeredReactionValueToAny
-} from "./modules/reactions/v1";
+import { MsgSaveProfileEncodeObject, MsgSaveProfileTypeUrl } from "./modules/profiles/v3";
 
 
 describe("DesmosClient", () => {
@@ -163,256 +113,6 @@ describe("DesmosClient", () => {
       const authInfo = AuthInfo.decode(signDoc.authInfoBytes);
       expect(authInfo.fee?.gasLimit).toBeDefined();
       expect(authInfo.fee?.amount).toHaveLength(1);
-    });
-
-    it("test MsgLinkChainAccount", async () => {
-      // Setup the client associated to the external wallet
-      const externalSigner = await OfflineSignerAdapter.fromMnemonic(
-        SigningMode.AMINO,
-        testUser2.mnemonic
-      );
-      const externalClient = await DesmosClient.connectWithSigner(
-        TEST_CHAIN_URL,
-        externalSigner
-      );
-      const externalAccounts = await externalSigner.getAccounts();
-      const externalAddress = externalAccounts[0].address;
-
-      // Setup the client associated to the Desmos profile
-      const [profileSigner, profileClient] = await getDirectSignerAndClient();
-      const profileAccounts = await profileSigner.getAccounts();
-      const profileAddress = profileAccounts[0].address;
-
-      // Create a profile (required to simulate the transaction properly)
-      const msgSaveProfile: MsgSaveProfileEncodeObject = {
-        typeUrl: "/desmos.profiles.v3.MsgSaveProfile",
-        value: MsgSaveProfile.fromPartial({
-          dtag: "JohnDoe",
-          creator: profileAddress,
-        }),
-      };
-      const saveProfileResult = await profileClient.signAndBroadcast(
-        profileAddress,
-        [msgSaveProfile],
-        "auto"
-      );
-      expect(saveProfileResult.code).toBe(0);
-
-      // Sign a dummy tx that includes the Desmos profile address within its memo, using the external account
-      const dummySignatureResult = await externalClient.signTx(
-        externalAddress,
-        [],
-        { fee: { amount: [], gas: "0" }, memo: profileAddress }
-      );
-
-      // Build the chain config
-      const chainConfig: ChainConfig = {
-        name: "desmos",
-      };
-
-      // Build the address data
-      const chainAddress: Any = bech32AddressToAny(
-        Bech32Address.fromPartial({
-          value: externalAddress,
-          prefix: "desmos",
-        })
-      );
-
-      // Build the signature
-      const signature: SingleSignature = SingleSignature.fromPartial({
-        valueType: SignatureValueType.SIGNATURE_VALUE_TYPE_COSMOS_AMINO, // Proper signature type
-        signature: getSignatureBytes(dummySignatureResult),
-      });
-
-      // Build the proof
-      const proof: Proof = Proof.fromPartial({
-        pubKey: dummySignatureResult.pubKey,
-        signature: singleSignatureToAny(signature),
-        plainText: toHex(getSignedBytes(dummySignatureResult)),
-      });
-
-      // Create the message to link the chain account
-      const msg: MsgLinkChainAccountEncodeObject = {
-        typeUrl: "/desmos.profiles.v3.MsgLinkChainAccount",
-        value: MsgLinkChainAccount.fromPartial({
-          chainConfig,
-          chainAddress,
-          proof,
-          signer: profileAddress,
-        }),
-      };
-
-      const result = await profileClient.signTx(profileAddress, [msg]);
-      expect(result.txRaw.signatures).toHaveLength(1);
-    });
-
-    it("test MsgCreatePost", async () => {
-      const [signer, client] = await getAminoSignerAndClient();
-      const { address } = (await signer.getAccounts())[0];
-
-      // Create a profile (required to create a post)
-      const msgSaveProfile: MsgSaveProfileEncodeObject = {
-        typeUrl: MsgSaveProfileTypeUrl,
-        value: {
-          dtag: "TestUser",
-          nickname: "Test user",
-          bio: DoNotModify,
-          profilePicture: DoNotModify,
-          coverPicture: DoNotModify,
-          creator: address,
-        },
-      };
-      await client.signAndBroadcast(address, [msgSaveProfile], "auto");
-      await sleep(5000);
-
-      // Create a subspace
-      // Ignore any error if the subspace already exists
-      const msgCreateSubspace: MsgCreateSubspaceEncodeObject = {
-        typeUrl: MsgCreateSubspaceTypeUrl,
-        value: {
-          name: "Test Subspace",
-          description: "Test subspaces",
-          owner: address,
-          creator: address,
-        },
-      };
-      await client.signAndBroadcast(address, [msgCreateSubspace], "auto");
-      await sleep(5000);
-
-      // Create a first post
-      const msgCreatePost: MsgCreatePostEncodeObject = {
-        typeUrl: MsgCreatePostTypeUrl,
-        value: {
-          subspaceId: Long.fromNumber(1),
-          sectionId: 0,
-          externalId: "",
-          text: "This is a test post",
-          entities: {
-            hashtags: [],
-            mentions: [],
-            urls: [
-              {
-                displayUrl: "IPFS",
-                start: Long.fromNumber(4),
-                end: Long.fromNumber(5),
-                url: "http://scripta.network/logo.svg",
-              },
-            ],
-          },
-          tags: [],
-          attachments: [],
-          author: address,
-          conversationId: Long.fromNumber(0),
-          replySettings: 1,
-          referencedPosts: [],
-        },
-      };
-      await client.signAndBroadcast(address, [msgCreatePost], "auto");
-      await sleep(5000);
-
-      // Create a post
-      const msg: MsgCreatePostEncodeObject = {
-        typeUrl: MsgCreatePostTypeUrl,
-        value: {
-          subspaceId: Long.fromNumber(1),
-          sectionId: 0,
-          externalId: "",
-          text: "This is a test post",
-          entities: {
-            hashtags: [],
-            mentions: [],
-            urls: [
-              {
-                displayUrl: "IPFS",
-                start: Long.fromNumber(4),
-                end: Long.fromNumber(5),
-                url: "http://scripta.network/logo.svg",
-              },
-            ],
-          },
-          tags: [],
-          attachments: [],
-          author: address,
-          conversationId: Long.fromNumber(0),
-          replySettings: 1,
-          referencedPosts: [
-            {
-              position: Long.fromNumber(0),
-              postId: Long.fromNumber(1),
-              type: 1,
-            },
-          ],
-        },
-      };
-      const result = await client.signAndBroadcast(address, [msg], "auto");
-      expect(result.code).toBe(0);
-    });
-
-    it("test MsgCreateReport", async () => {
-      const [signer, client] = await getAminoSignerAndClient();
-      const { address } = (await signer.getAccounts())[0];
-
-      // Create a report reason
-      const msgAddReasons: MsgAddReasonEncodeObject = {
-        typeUrl: MsgAddReasonTypeUrl,
-        value: {
-          signer: address,
-          subspaceId: Long.fromNumber(1),
-          description: "Test reason",
-          title: "Test reason",
-        },
-      };
-      await client.signAndBroadcast(address, [msgAddReasons], "auto");
-      await sleep(5000);
-
-      // report a post
-      const msgCreateReport: MsgCreateReportEncodeObject = {
-        typeUrl: MsgCreateReportTypeUrl,
-        value: {
-          subspaceId: Long.fromNumber(1),
-          message: "This is a test report",
-          reasonsIds: [1],
-          reporter: address,
-          target: postTargetToAny({
-            postId: Long.fromNumber(1),
-          }),
-        },
-      };
-      await client.signAndBroadcast(address, [msgCreateReport], "auto");
-      await sleep(5000);
-    });
-
-    it("test MsgAddReaction", async () => {
-      const [signer, client] = await getAminoSignerAndClient();
-      const { address } = (await signer.getAccounts())[0];
-
-      // Create a registered reaction for the subspace
-      const msgRegisterReaction: MsgAddRegisteredReactionEncodeObject = {
-        typeUrl: MsgAddRegisteredReactionTypeUrl,
-        value: {
-          user: address,
-          subspaceId: Long.fromNumber(1),
-          displayValue: "like",
-          shorthandCode: "like",
-        },
-      };
-      await client.signAndBroadcast(address, [msgRegisterReaction], "auto");
-      await sleep(5000);
-
-      // React to a post
-      const msgAddReaction: MsgAddReactionEncodeObject = {
-        typeUrl: MsgAddReactionTypeUrl,
-        value: {
-          subspaceId: Long.fromNumber(1),
-          postId: Long.fromNumber(1),
-          user: address,
-          value: registeredReactionValueToAny({
-            registeredReactionId: 1,
-          }),
-        },
-      };
-      await client.signAndBroadcast(address, [msgAddReaction], "auto");
-      await sleep(5000);
     });
   });
